requirements.txt
Pillow
numpy
matplotlib
tensorflow
